nginx: forcibly close HTTP/3 stream when finalizing QUIC connection
Due to the fact that the underlying QUIC connection is already closed
(e.g. due to idle timeout), the open HTTP/3 streams need to be forcibly
terminated (e.g. if a stream was blocked, it would end up never being
unblocked because the client can't provide additional stream credit, so
the stream would never be closed).